% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/career_paths.R
\name{bfs_find_paths}
\alias{bfs_find_paths}
\title{BFS to find all paths between two nodes}
\usage{
bfs_find_paths(adj_list, start, end, max_depth)
}
\arguments{
\item{adj_list}{Adjacency list}

\item{start}{Start node}

\item{end}{End node}

\item{max_depth}{Maximum path length}
}
\value{
List of paths (each path is a character vector)
}
\description{
BFS to find all paths between two nodes
}
\keyword{internal}
