% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/career_paths.R
\name{build_adjacency_list}
\alias{build_adjacency_list}
\title{Build adjacency list from edges}
\usage{
build_adjacency_list(edges, from_col, to_col)
}
\arguments{
\item{edges}{Edge data frame}

\item{from_col}{Name of source column}

\item{to_col}{Name of target column}
}
\value{
Named list where each element contains outgoing neighbors
}
\description{
Build adjacency list from edges
}
\keyword{internal}
