% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdesc_char}
\alias{cdesc_char}
\title{An example table of metadata, as would be
parsed from or parse.gctx. Initially all the
columns are of type character.}
\format{
An object of class \code{data.frame} with 368 rows and 8 columns.
}
\usage{
cdesc_char
}
\description{
An example table of metadata, as would be
parsed from or parse.gctx. Initially all the
columns are of type character.
}
\keyword{datasets}
