% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_dups}
\alias{check_dups}
\title{Check for duplicates in a vector}
\usage{
check_dups(x, name = "")
}
\arguments{
\item{x}{the vector}

\item{name}{the name of the object to print
in an error message if duplicates are found}
}
\value{
silently returns NULL
}
\description{
Check for duplicates in a vector
}
\examples{
check_dups(c("a", "b", "c", "a", "d"))
}
