% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cmap_log}
\alias{cmap_log}
\title{Standardized CLI Logging}
\usage{
cmap_log(level = c("info", "success", "warn"), message, ...)
}
\arguments{
\item{level}{Character string specifying the log level: "info", "success", or "warn".}

\item{message}{Character string with the message to log.}

\item{...}{Additional arguments passed to the cli function.}
}
\description{
Provides consistent logging messages using cli functions.
}
