% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract.gct}
\alias{extract.gct}
\title{Exract elements from a GCT matrix}
\usage{
extract.gct(g, row_field, col_field, rdesc = NULL, cdesc = NULL)
}
\arguments{
\item{g}{the GCT object}

\item{row_field}{the column name in rdesc to search on}

\item{col_field}{the column name in cdesc to search on}

\item{rdesc}{a \code{data.frame} of row annotations}

\item{cdesc}{a \code{data.frame} of column annotations}
}
\value{
a list of the following elements
* \code{mask} a logical matrix of the same dimensions as
      \code{ds@mat} indicating which matrix elements have
      been extracted
* \code{idx} a numerical vector index into \code{ds@mat}
      representing which elements have been extracted
* \code{vals} a vector of the extracted values
}
\description{
extract the elements from a \code{GCT} object
  where the values of \code{row_field} and \code{col_field}
  are the same. A concrete example is if \code{ds} represents
  a pairwise similarity matrix between perturbagens and you
  want to extract all the similarities where the row and 
  column perturbagens are the same.
}
\examples{
m <- cor(ds@mat)
g <- new("GCT", mat=m, rdesc=ds@cdesc, cdesc=ds@cdesc)
same_pert_cor <- extract.gct(g, row_field="pert_id", col_field="pert_id")
str(same_pert_cor)

}
