% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_csv_files}
\alias{find_csv_files}
\title{Helper: List all CSV files recursively}
\usage{
find_csv_files(dir_path)
}
\arguments{
\item{dir_path}{Path to directory}
}
\value{
Character vector of file paths
}
\description{
Helper: List all CSV files recursively
}
