% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{fix_datatypes}
\alias{fix_datatypes}
\title{Adjust the data types for columns of a meta data frame}
\usage{
fix_datatypes(meta)
}
\arguments{
\item{meta}{a data.frame}
}
\value{
meta the same data frame with (potentially) adjusted
  column types.
}
\description{
GCT(X) parsing initially returns data frames
  of row and column descriptors where all columns are of
  type character. This is inconvenient for analysis, so
  the goal of this function is to try and guess the
  appropriate data type for each column.
}
\details{
This is a low-level helper function
  which most users will not need to access directly
}
\examples{
# meta data table with all character types
str(cdesc_char)
fixed <- cmapR:::fix_datatypes(cdesc_char)
# note how some column classes have changed
str(fixed)

}
\seealso{
Other GCTX parsing functions: 
\code{\link{append_dim}()},
\code{\link{parse.gctx}()},
\code{\link{process_ids}()},
\code{\link{read_gctx_ids}()},
\code{\link{read_gctx_meta}()},
\code{\link{write_gctx_meta}()},
\code{\link{write_gctx}()},
\code{\link{write_gct}()}
}
\concept{GCTX parsing functions}
\keyword{internal}
