% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_gct_ids}
\alias{get_gct_ids}
\alias{get_gct_ids,GCT-method}
\title{Extract the ids from a GCT object}
\usage{
get_gct_ids(g, dim = "row")

\S4method{get_gct_ids}{GCT}(g, dim = "row")
}
\arguments{
\item{g}{the GCT object}

\item{dim}{which dimension to extract}
}
\value{
a vector of ids
}
\description{
Extract the ids from a GCT object
}
\examples{
m <- get_gct_ids(ds)
}
\seealso{
Other GCT accessor methods: 
\code{\link{get_gct_matrix}()},
\code{\link{get_gct_meta}()}
}
\concept{GCT accessor methods}
