% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_gct_matrix}
\alias{get_gct_matrix}
\alias{get_gct_matrix,GCT-method}
\title{Extract the matrix from a GCT object}
\usage{
get_gct_matrix(g)

\S4method{get_gct_matrix}{GCT}(g)
}
\arguments{
\item{g}{the GCT object}
}
\value{
a matrix
}
\description{
Extract the matrix from a GCT object
}
\examples{
m <- get_gct_matrix(ds)
}
\seealso{
Other GCT accessor methods: 
\code{\link{get_gct_ids}()},
\code{\link{get_gct_meta}()}
}
\concept{GCT accessor methods}
