% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_gct_meta}
\alias{get_gct_meta}
\alias{get_gct_meta,GCT-method}
\title{Extract the metadata from a GCT object}
\usage{
get_gct_meta(g, dim = "row")

\S4method{get_gct_meta}{GCT}(g, dim = "row")
}
\arguments{
\item{g}{the GCT object}

\item{dim}{which dimension to extract}
}
\value{
a data.frame of metadata
}
\description{
Extract the metadata from a GCT object
}
\examples{
m <- get_gct_meta(ds)
}
\seealso{
Other GCT accessor methods: 
\code{\link{get_gct_ids}()},
\code{\link{get_gct_matrix}()}
}
\concept{GCT accessor methods}
