% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCT.R
\name{ids}
\alias{ids}
\alias{ids,GCT-method}
\alias{ids<-,GCT-method}
\title{Extract the or set row or column ids of a GCT object}
\usage{
ids(g, dimension = "row")

\S4method{ids}{GCT}(g, dimension = "row")

\S4method{ids}{GCT}(g, dimension = "row") <- value
}
\arguments{
\item{g}{the GCT object}
}
\value{
a vector of row ids
}
\description{
Extract the or set row or column ids of a GCT object
}
\examples{
# extract rids
rids <- ids(ds)
# extract column ids
cids <- ids(ds, "column")
# set rids
ids(ds) <- as.character(1:length(rid))
# set cids
ids(ds, "column") <- as.character(1:length(cid))
}
\seealso{
Other GCT accessor methods: 
\code{\link{mat}()},
\code{\link{meta}()}
}
\concept{GCT accessor methods}
