% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{Check if x is a whole number}
\usage{
is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{number to test}

\item{tol}{the allowed tolerance}
}
\value{
boolean indicating whether x is tol away from a whole number value
}
\description{
Check if x is a whole number
}
\examples{
is.wholenumber(1)
is.wholenumber(0.5)
}
