% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kd_gct}
\alias{kd_gct}
\title{An example GCT object of knockdown experiments
targeting a subset of landmark genes.}
\format{An object of class \code{GCT} with 976 rows and 374 columns.}
\usage{
kd_gct
}
\description{
An example GCT object of knockdown experiments
targeting a subset of landmark genes.
}
\keyword{datasets}
