% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCT.R
\name{mat}
\alias{mat}
\alias{mat,GCT-method}
\alias{mat<-}
\alias{mat<-,GCT-method}
\title{Extract or set the matrix of GCT object}
\usage{
mat(g)

\S4method{mat}{GCT}(g)

mat(g) <- value

\S4method{mat}{GCT}(g) <- value
}
\arguments{
\item{g}{the GCT object}

\item{value}{a numeric matrix}
}
\value{
a matrix
}
\description{
Extract or set the matrix of GCT object
}
\examples{
# get the matrix
m <- mat(ds)
# set the matrix
mat(ds) <- matrix(0, nrow=nrow(m), ncol=ncol(m))
}
\seealso{
Other GCT accessor methods: 
\code{\link{ids}()},
\code{\link{meta}()}
}
\concept{GCT accessor methods}
