% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/career_paths.R
\name{match_title}
\alias{match_title}
\title{Match a title string to available titles}
\usage{
match_title(query, titles, ignore_case = TRUE)
}
\arguments{
\item{query}{The search query}

\item{titles}{Vector of available titles}

\item{ignore_case}{Whether to ignore case}
}
\value{
Matched title or NA
}
\description{
Match a title string to available titles
}
\keyword{internal}
