% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge.gct}
\alias{merge.gct}
\title{Merge two GCT objects together}
\usage{
merge.gct(g1, g2, dimension = "row", matrix_only = F)
}
\arguments{
\item{g1}{the first GCT object}

\item{g2}{the second GCT object}

\item{dimension}{the dimension on which to merge (row or column)}

\item{matrix_only}{boolean idicating whether to keep only the
data matrices from \code{g1} and \code{g2} and ignore their
row and column meta data}
}
\description{
Merge two GCT objects together
}
\examples{
# take the first 10 and last 10 rows of an object
# and merge them back together
(a <- subset.gct(ds, rid=1:10))
(b <- subset.gct(ds, rid=969:978))
(merged <- merge.gct(a, b, dimension="row"))

}
\seealso{
Other GCT utilities: \code{\link{annotate.gct}},
  \code{\link{melt.gct}}, \code{\link{rank.gct}},
  \code{\link{subset.gct}}
}
