% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{na_pad_matrix}
\alias{na_pad_matrix}
\title{Pad a matrix with additional rows/columns of NA values}
\usage{
na_pad_matrix(m, row_universe = NULL, col_universe = NULL)
}
\arguments{
\item{m}{a matrix with unique row and column names}

\item{row_universe}{a vector with the universe of possible
row names}

\item{col_universe}{a vector with the universe of possible
column names}
}
\value{
a matrix
}
\description{
Pad a matrix with additional rows/columns of NA values
}
\examples{
m <- matrix(rnorm(10), nrow=2)
rownames(m) <- c("A", "B")
colnames(m) <- letters[1:5]
na_pad_matrix(m, row_universe=LETTERS, col_universe=letters)

}
