% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{parse.gmx}
\alias{parse.gmx}
\title{Read a GMX file and return a list}
\usage{
parse.gmx(fname)
}
\arguments{
\item{fname}{the file path to be parsed}
}
\value{
a list of the contents of \code{fname}. See details.
}
\description{
Read a GMX file and return a list
}
\details{
\code{parse.gmx} returns a nested list object. The top
  level contains one list per column in \code{fname}. Each of 
  these is itself a list with the following fields:
  - \code{head}: the name of the data (column in \code{fname})
  - \code{desc}: description of the corresponding data
  - \code{len}: the number of data items
  - \code{entry}: a vector of the data items
}
\examples{
gmx_path <- system.file("extdata", "lm_probes.gmx", package="cmapR")
gmx <- parse.gmx(gmx_path)
str(gmx)

}
\seealso{
\link{http://clue.io/help} for details on the GMX file format

Other CMap parsing functions: \code{\link{parse.gmt}},
  \code{\link{parse.grp}}, \code{\link{write.gmt}},
  \code{\link{write.grp}}
}
