% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/io.R
\name{parse.grp}
\alias{parse.grp}
\alias{parse_grp}
\title{Read a GRP file and return a vector of its contents}
\usage{
parse.grp(...)

parse_grp(fname)
}
\arguments{
\item{fname}{the file path to be parsed}
}
\value{
a vector of the contents of \code{fname}
}
\description{
Read a GRP file and return a vector of its contents
}
\examples{
grp_path <- system.file("extdata", "lm_epsilon_n978.grp", package="cmapR")
values <- parse_grp(grp_path)
str(values)
}
\seealso{
Visit \url{http://clue.io/help} for details on the GRP file format

Other CMap parsing functions: 
\code{\link{lxb2mat}()},
\code{\link{parse.gmt}()},
\code{\link{parse.gmx}()},
\code{\link{write_gmt}()},
\code{\link{write_grp}()}
}
\concept{CMap parsing functions}
