% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read.gctx.ids}
\alias{read.gctx.ids}
\title{Read GCTX row or column ids}
\usage{
read.gctx.ids(gctx_path, dimension = "row")
}
\arguments{
\item{gctx_path}{path to the GCTX file}

\item{dimension}{which ids to read (row or column)}
}
\value{
a character vector of row or column ids from the provided file
}
\description{
Read GCTX row or column ids
}
\examples{
gct_file <- system.file("extdata", "modzs_n272x978.gctx", package="cmapR")
# row ids
rid <- read.gctx.ids(gct_file)
head(rid)
# column ids
cid <- read.gctx.ids(gct_file, dimension="column")
head(cid)

}
\seealso{
Other GCTX parsing functions: \code{\link{append.dim}},
  \code{\link{fix.datatypes}}, \code{\link{parse.gctx}},
  \code{\link{process_ids}}, \code{\link{read.gctx.meta}},
  \code{\link{write.gctx.meta}}, \code{\link{write.gctx}},
  \code{\link{write.gct}}
}

