% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset.gct}
\alias{subset.gct}
\title{Subset a gct object using the provided row and column ids}
\usage{
subset.gct(g, rid = NULL, cid = NULL)
}
\arguments{
\item{g}{a gct object}

\item{rid}{a vector of character ids or integer indices for ROWS}

\item{cid}{a vector of character ids or integer indices for COLUMNS}
}
\description{
Subset a gct object using the provided row and column ids
}
\examples{
# first 10 rows and columns by index
(a <- subset.gct(ds, rid=1:10, cid=1:10))

# first 10 rows and columns using character ids
(b <- subset.gct(ds, rid=ds@rid[1:10], cid=ds@cid[1:10]))

identical(a, b) # TRUE

}
\seealso{
Other GCT utilities: \code{\link{melt.gct}},
  \code{\link{merge.gct}}
}

