% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset_to_ids}
\alias{subset_to_ids}
\title{Do a robust \code{\link{data.frame}} subset to a set of ids}
\usage{
subset_to_ids(df, ids)
}
\arguments{
\item{df}{\code{\link{data.frame}} to subset}

\item{ids}{the ids to subset to}
}
\value{
a subset version of \code{df}
}
\description{
Do a robust \code{\link{data.frame}} subset to a set of ids
}
\keyword{internal}

