% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{threshold}
\alias{threshold}
\title{Threshold a numeric vector}
\usage{
threshold(x, minval, maxval)
}
\arguments{
\item{x}{the vector}

\item{minval}{minium allowed value}

\item{maxval}{maximum allowed value}
}
\value{
a thresholded version of \code{x}
}
\description{
Threshold a numeric vector
}
\examples{
x <- rnorm(20)
threshold(x, -0.1, -0.1)

}
