% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transpose.gct}
\alias{transpose.gct}
\title{Transpose a GCT object}
\usage{
transpose.gct(g)
}
\arguments{
\item{g}{the \code{GCT} object}
}
\value{
a modified verion of the input \code{GCT} object
  where the matrix has been transposed and the row and column
  ids and annotations have been swapped.
}
\description{
Transpose a GCT object
}
\examples{
transpose.gct(ds)

}
