% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/utils.R
\name{transpose.gct}
\alias{transpose.gct}
\alias{transpose_gct}
\alias{transpose_gct,GCT-method}
\title{Transpose a GCT object}
\usage{
transpose.gct(...)

transpose_gct(g)

\S4method{transpose_gct}{GCT}(g)
}
\arguments{
\item{...}{arguments passed on to \code{transpose_gct}}

\item{g}{the \code{GCT} object}
}
\value{
a modified verion of the input \code{GCT} object
  where the matrix has been transposed and the row and column
  ids and annotations have been swapped.
}
\description{
Transpose a GCT object
}
\examples{
transpose_gct(ds)

}
\concept{GCT utilties}
