% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write.gctx}
\alias{write.gctx}
\title{Write a GCT object to disk in GCTX format}
\usage{
write.gctx(ds, ofile, appenddim = T, compression_level = 0,
  matrix_only = F)
}
\arguments{
\item{ds}{a GCT object}

\item{ofile}{the desired file path for writing}

\item{appenddim}{boolean indicating whether the
resulting filename will have dimensions appended
(e.g. my_file_n384x978.gctx)}

\item{compression_level}{integer between 1-9 indicating
how much to compress data before writing. Higher values
result in smaller files but slower read times.}

\item{matrix_only}{boolean indicating whether to write
only the matrix data (and skip row, column annotations)}
}
\description{
Write a GCT object to disk in GCTX format
}
\examples{
\dontrun{
# assume ds is a GCT object
write.gctx(ds, "my/desired/outpath/and/filename")
}
}
\seealso{
Other GCTX parsing functions: \code{\link{append.dim}},
  \code{\link{fix.datatypes}}, \code{\link{parse.gctx}},
  \code{\link{process_ids}}, \code{\link{read.gctx.ids}},
  \code{\link{read.gctx.meta}}, \code{\link{write.gct}},
  \code{\link{write.meta}}
}

