% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write.gmt}
\alias{write.gmt}
\title{Write a nested list to a GMT file}
\usage{
write.gmt(lst, fname)
}
\arguments{
\item{lst}{the nested list to write. See \code{details}.}

\item{fname}{the desired file name}
}
\description{
Write a nested list to a GMT file
}
\details{
\code{lst} needs to be a nested list where each 
  sub-list is itself a list with the following fields:
  - \code{head}: the name of the data
  - \code{desc}: description of the corresponding data
  - \code{len}: the number of data items
  - \code{entry}: a vector of the data items
}
\examples{
\dontrun{
write.gmt(gene_set, "gene_set.gmt")
}

}
\seealso{
\link{http://clue.io/help} for details on the GMT file format

Other CMap parsing functions: \code{\link{parse.gmt}},
  \code{\link{parse.gmx}}, \code{\link{parse.grp}},
  \code{\link{write.grp}}
}

