% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_gctx}
\alias{write_gctx}
\title{Write a GCT object to disk in GCTX format}
\usage{
write_gctx(
  ds,
  ofile,
  appenddim = TRUE,
  compression_level = 0,
  matrix_only = FALSE,
  max_chunk_kb = 1024
)
}
\arguments{
\item{ds}{a GCT object}

\item{ofile}{the desired file path for writing}

\item{appenddim}{boolean indicating whether the
resulting filename will have dimensions appended
(e.g. my_file_n384x978.gctx)}

\item{compression_level}{integer between 1-9 indicating
how much to compress data before writing. Higher values
result in smaller files but slower read times.}

\item{matrix_only}{boolean indicating whether to write
only the matrix data (and skip row, column annotations)}

\item{max_chunk_kb}{for chunking, the maximum number of KB
a given chunk will occupy}
}
\value{
silently returns NULL
}
\description{
Write a GCT object to disk in GCTX format
}
\examples{
# note this will create a GCT file in your current directory
write_gctx(ds, "dataset")

}
\seealso{
Other GCTX parsing functions: 
\code{\link{append_dim}()},
\code{\link{fix_datatypes}()},
\code{\link{parse.gctx}()},
\code{\link{process_ids}()},
\code{\link{read_gctx_ids}()},
\code{\link{read_gctx_meta}()},
\code{\link{write_gctx_meta}()},
\code{\link{write_gct}()}
}
\concept{GCTX parsing functions}
