% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_grp}
\alias{write_grp}
\title{Write a vector to a GRP file}
\usage{
write_grp(vals, fname)
}
\arguments{
\item{vals}{the vector of values to be written}

\item{fname}{the desired file name}
}
\value{
silently returns NULL
}
\description{
Write a vector to a GRP file
}
\examples{
\dontrun{
write_grp(letters, "letter.grp")
}

}
\seealso{
Visit \url{http://clue.io/help} for details on the GRP file format

Other CMap parsing functions: 
\code{\link{lxb2mat}()},
\code{\link{parse.gmt}()},
\code{\link{parse.gmx}()},
\code{\link{parse.grp}()},
\code{\link{write_gmt}()}
}
\concept{CMap parsing functions}
