% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_tbl}
\alias{write_tbl}
\title{Write a \code{data.frame} to a tab-delimited text file}
\usage{
write_tbl(tbl, ofile, ...)
}
\arguments{
\item{tbl}{the \code{data.frame} to be written}

\item{ofile}{the desired file name}

\item{...}{additional arguments passed on to \code{write.table}}
}
\value{
silently returns NULL
}
\description{
Write a \code{data.frame} to a tab-delimited text file
}
\details{
This method simply calls \code{write.table} with some
  preset arguments that generate a unquoated, tab-delimited file
  without row names.
}
\examples{
\dontrun{
write_tbl(cdesc_char, "col_meta.txt")
}

}
\seealso{
\code{\link{write.table}}
}
