% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ds}
\alias{ds}
\title{An example of a GCT object with row and
column metadata and gene expression values
in the matrix.}
\format{An object of class \code{GCT} of length 1.}
\usage{
ds
}
\description{
An example of a GCT object with row and
column metadata and gene expression values
in the matrix.
}
\keyword{datasets}

