% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_gctx_ids}
\alias{read_gctx_ids}
\title{Read GCTX row or column ids}
\usage{
read_gctx_ids(gctx_path, dim = "row")
}
\arguments{
\item{gctx_path}{path to the GCTX file}

\item{dim}{which ids to read (row or column)}
}
\value{
a character vector of row or column ids from the provided file
}
\description{
Read GCTX row or column ids
}
\examples{
gct_file <- system.file("extdata", "modzs_n25x50.gctx", package="cmapR")
# row ids
rid <- read_gctx_ids(gct_file)
head(rid)
# column ids
cid <- read_gctx_ids(gct_file, dim="column")
head(cid)

}
\seealso{
Other GCTX parsing functions: 
\code{\link{append_dim}()},
\code{\link{fix_datatypes}()},
\code{\link{parse_gctx}()},
\code{\link{process_ids}()},
\code{\link{read_gctx_meta}()},
\code{\link{write_gctx_meta}()},
\code{\link{write_gctx}()},
\code{\link{write_gct}()}
}
\concept{GCTX parsing functions}
