% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_gctx_meta}
\alias{write_gctx_meta}
\title{Write a \code{data.frame} of meta data to GCTX file}
\usage{
write_gctx_meta(ofile, df, dimension = "row")
}
\arguments{
\item{ofile}{the desired file path for writing}

\item{df}{the \code{data.frame} of annotations}

\item{dimension}{the dimension to annotate
(row or column)}
}
\value{
silently returns NULL
}
\description{
Write a \code{data.frame} of meta data to GCTX file
}
\examples{
\dontrun{
# assume ds is a GCT object
write_gctx_meta("/my/file/path", cdesc_char, dimension="col")
}
}
\seealso{
Other GCTX parsing functions: 
\code{\link{append_dim}()},
\code{\link{fix_datatypes}()},
\code{\link{parse.gctx}()},
\code{\link{process_ids}()},
\code{\link{read_gctx_ids}()},
\code{\link{read_gctx_meta}()},
\code{\link{write_gctx}()},
\code{\link{write_gct}()}
}
\concept{GCTX parsing functions}
\keyword{internal}
